+++ grade.h
@@ -1,0 +1,10 @@
+#include "median.h"
+#include "student_info.h"
+#include <string>
+#include <vector>
+
+
+
+double grade (double midterm, double final, double homework);
+double grade (double midterm, double final, const vector<double>& hw) ;
+double grade (const Student_info& s);
